+++ src/features/common/WebappButton/WebappButton.module.scss
@@ -49,7 +49,7 @@
   font-weight: 700;
   color: #2f3336;
   cursor: pointer;
-  line-height: 1.2px;
+  line-height: 16px;
   text-wrap: wrap;
 }
