+++ src/features/projectsV2/ProjectSnippet/styles/Badge.module.scss
@@ -38,7 +38,3 @@
   font-weight: 400;
   color: $primaryColor;
 }
-
-.topProjectIcon {
-  margin-bottom: 3.5px;
-}
